heroes - crud terminado

+++ src/app/heroes/pages/agregar/agregar.component.ts
@@ -1,6 +1,9 @@
 import { Component, OnInit } from '@angular/core';
+import { MatDialog } from '@angular/material/dialog';
+import { MatSnackBar } from '@angular/material/snack-bar';
 import { ActivatedRoute, Router } from '@angular/router';
 import { switchMap } from 'rxjs';
+import { ConfirmarComponent } from '../../components/confirmar/confirmar.component';
 
 import { Heroe, Publisher } from '../../interfaces/heroes.interface';
 import { HeroesService } from '../../services/heroes.service';
@@ -45,7 +48,9 @@ export class AgregarComponent implements OnInit {
   constructor(
     private heroService: HeroesService,
     private activatedRoute: ActivatedRoute,
-    private router: Router
+    private router: Router,
+    private snackBar: MatSnackBar,
+    private dialog: MatDialog
   ) { }
 
   ngOnInit(): void {
@@ -71,21 +76,41 @@ export class AgregarComponent implements OnInit {
 
     if (this.heroe.id) {
       //? Actualzar
-      this.heroService.updateHero(this.heroe).subscribe(res => console.log('Actualizando: ', res));
+      this.heroService.updateHero(this.heroe).subscribe(res => this.showSnakBar('Registro actualizado'));
     } else {
       //? Insertar
 
       this.heroService.addHero(this.heroe).subscribe(heroe => {
-        this.router.navigate(['/heroes/editar', heroe.id])
+        this.router.navigate(['/heroes/editar', heroe.id]);
+        this.showSnakBar('Registro creado');
       });
-    }    
+    }
   }
 
   delete() {
-    this.heroService.deleteHero( this.heroe.id! )
-      .subscribe( resp =>  {
-        this.router.navigate(['/heroes'])
-      })
+    const dialog = this.dialog.open(ConfirmarComponent, {
+      width: '250px',
+      data: { ...this.heroe }
+    });
+
+    dialog.afterClosed().subscribe(
+      (result) => {
+        if (result){
+          this.heroService.deleteHero(this.heroe.id!)
+            .subscribe(resp => {
+              this.router.navigate(['/heroes']);
+              this.showSnakBar('Registro eliminado')
+            }
+          )
+        }
+      }
+    )
+
+    
+  }
+
+  showSnakBar(msg: string) {
+    this.snackBar.open(msg, 'Ok!', { duration: 2500 });
   }
 
 }
